requirements.txt
markdownify==0.12.1
beautifulsoup4==4.12.3
pyautogen==0.2.26
autogenbench==0.0.3
